Moved Battle and BattleLocation to subpackage
 Changes to be committed:
	modified:   src/main/java/com/ardaslegends/domain/war/War.java
	renamed:    src/main/java/com/ardaslegends/domain/war/Battle.java -> src/main/java/com/ardaslegends/domain/war/battle/Battle.java
	renamed:    src/main/java/com/ardaslegends/domain/war/BattleLocation.java -> src/main/java/com/ardaslegends/domain/war/battle/BattleLocation.java
	modified:   src/main/java/com/ardaslegends/presentation/api/BattleRestController.java
	modified:   src/main/java/com/ardaslegends/presentation/api/response/war/BattleResponse.java
	modified:   src/main/java/com/ardaslegends/repository/BattleRepository.java
	modified:   src/main/java/com/ardaslegends/service/war/BattleService.java
	modified:   src/test/java/com/ardaslegends/repository/ArmyRepositoryTest.java
	modified:   src/test/java/com/ardaslegends/service/BattleServiceTest.java
