+++ src/main/java/com/ardaslegends/repository/BattleRepository.java
@@ -1,6 +1,6 @@
 package com.ardaslegends.repository;
 
-import com.ardaslegends.domain.war.Battle;
+import com.ardaslegends.domain.war.battle.Battle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BattleRepository extends JpaRepository<Battle, Long> {
