+++ src/main/java/com/ardaslegends/service/war/BattleService.java
@@ -1,8 +1,8 @@
 package com.ardaslegends.service.war;
 
 import com.ardaslegends.domain.*;
-import com.ardaslegends.domain.war.Battle;
-import com.ardaslegends.domain.war.BattleLocation;
+import com.ardaslegends.domain.war.battle.Battle;
+import com.ardaslegends.domain.war.battle.BattleLocation;
 import com.ardaslegends.repository.*;
 import com.ardaslegends.repository.war.WarRepository;
 import com.ardaslegends.repository.war.QueryWarStatus;
